@@ -2,19 +2,19 @@
   "name": "socialis",
   "version": "1.0.0",
   "dependencies": {
-    "@angular/animations": "^4.0.1",
+    "@angular/animations": "6.0.6",
     "@angular/cdk": "^2.0.0-beta.8",
-    "@angular/common": "^4.0.1",
-    "@angular/compiler": "^4.0.1",
-    "@angular/core": "^4.0.1",
-    "@angular/forms": "^4.0.1",
-    "@angular/http": "^4.0.1",
-    "@angular/platform-browser": "^4.0.0",
-    "@angular/platform-browser-dynamic": "^4.0.0",
-    "@angular/router": "^4.0.0",
+    "@angular/common": "6.0.6",
+    "@angular/compiler": "6.0.6",
+    "@angular/core": "6.0.6",
+    "@angular/forms": "6.0.6",
+    "@angular/http": "6.0.6",
+    "@angular/platform-browser": "6.0.6",
+    "@angular/platform-browser-dynamic": "6.0.6",
+    "@angular/router": "6.0.6",
     "@types/lodash": "^4.14.58",
     "@types/node": "^6.0.68",
-    "conichi-material": "^1.17.26",
+    "conichi-material": "^1.18.1",
     "cors": "^2.8.4",
     "dotenv": "^4.0.0",
     "express": "^4.14.1",
@@ -28,17 +28,20 @@
     "node-linked-in": "0.0.7",
     "node-mysql": "^0.4.2",
     "nodemailer": "^2.7.2",
+    "rxjs": "^6.2.1",
+    "rxjs-compat": "^6.2.1",
     "twitter": "^1.7.0"
   },
   "devDependencies": {
-    "@angular/cli": "1.0.0",
-    "@angular/compiler-cli": "^4.0.0",
+    "@angular/cli": "6.0.8",
+    "@angular/compiler-cli": "6.0.6",
     "@types/node": "^6.0.68",
     "codelyzer": "~2.0.0",
     "nodemon": "^1.11.0",
     "ts-node": "~2.0.0",
     "tslint": "~4.5.0",
-    "typescript": "~2.2.0"
+    "typescript": "2.7.2",
+    "@angular-devkit/build-angular": "~0.6.8"
   },
   "engines": {
     "node": ">=0.10.0"
