@@ -7,25 +7,27 @@ const User = require("../models/users");
 router.post("/login", (req, res, next) => {
   const id = req.body.productId;
   
-  User.findOne({ email: req.body.email, password: req.body.password })
+  User.findOne({ email: req.body.user.email, password: req.body.user.password }, 'first_name last_name email')
     .exec()
-    .then(doc => {
-      if (doc) 
-        res.status(200).json(doc);
+    .then(user => {
+      if (user)
+        res.status(200).json({ user });
       else
-        res.status(404).json({ message: 'No valid entry found for provided ID' });
+        res.status(400).json({ message: 'No valid entry found for provided ID' });
     })
     .catch(err => res.status(500).json({ error: err }));
 });
 
 router.post("/register", (req, res, next) => {
-  req.body.id = new mongoose.Types.ObjectId()
-  
-  const user = new User(req.body);
+  req.body.user.id = new mongoose.Types.ObjectId();
+  req.body.user.profile_photo = './assets/images/default-user.jpg';
+  req.body.user.name = `${ req.body.user.first_name } ${ req.body.user.last_name }`;
+    
+  const user = new User(req.body.user);
   
   user
     .save()
-    .then(result => res.status(201).json(result))
+    .then(user => res.status(201).json({ user }))
     .catch(err => res.status(500).json({ error: err }));
 });
 
