@@ -1,14 +1,13 @@
 // Modules
-const keywords = require("./keywords")
+const keywords = require("./keywords");
 
 module.exports = (title) => {
-  const wordArray = title.split(' ')
+  const wordArray = title.split(' ');
 
   for (i = 0; i < wordArray.length; ++i) {
-    if (keywords.indexOf(wordArray[i].toLowerCase()) > -1) {
-      wordArray[i] = '#' + wordArray[i]
-    }
-  }
+    if (keywords.indexOf(wordArray[i].toLowerCase()) > -1)
+      wordArray[i] = '#' + wordArray[i];
+  };
 
-  return wordArray.join(" ")
+  return wordArray.join(" ");
 };
