@@ -1,22 +1,20 @@
 // Node modules
 const nodemailer = require('nodemailer');
 
-
-module.exports = function(message, error) {
-  console.log(message, error);
-  
-  const transporter = nodemailer.createTransport('smtps://' + process.env.EMAIL_ADDRESS + ':' + process.env.EMAIL_PASS + '@' + process.env.EMAIL_HOST);
+module.exports = (message, error) => {  
+  const transporter = nodemailer.createTransport(
+    `smtps://${ process.env.EMAIL_ADDRESS }:${ process.env.EMAIL_PASS }@${ process.env.EMAIL_HOST }`
+  );
 
   const emailTemplate = {
     from: process.env.EMAIL_ADDRESS,
     to: process.env.EMAIL_ADDRESS,
     subject: 'Error!!',
-    text: message + ' ' + JSON.stringify(error),
-    html: message + '<br><br>' + JSON.stringify(error)
+    text: `${ message } ${ JSON.stringify(error) }`,
+    html: `${ message }<br><br>${ JSON.stringify(error) }`
   };
   
-  transporter.sendMail(emailTemplate, function(error, info){
-    if(error){ return console.log(error); }
-    console.log('Sent error mail');
+  transporter.sendMail(emailTemplate, (error, info) => {
+    if(error) return console.log('Sent error mail', error);
   });
 };
