@@ -1,4 +1,4 @@
-module.exports = function(max) {
+module.exports = (max) => {
   var rand = Math.random() * (max - 0);
   return Math.floor(rand);
 };
