@@ -1,27 +1,23 @@
 // Node modules
-var Twitter = require('twitter');
+const Twitter = require('twitter');
 
 // Modules
-var mailError = require('./emailAlerts');
+const mailError = require('./emailAlerts');
 
 // Config
 require('dotenv').config();
 
 
 exports.post = function(body) {
-  var client = new Twitter({
+  const client = new Twitter({
     consumer_key: process.env.TWITTER_CONSUMER_KEY,
     consumer_secret: process.env.TWITTER_CONSUMER_SECRET,
     access_token_key: process.env.TWITTER_TOKEN_KEY,
     access_token_secret: process.env.TWITTER_TOKEN_SECRET
   });
 
-  client.post('statuses/update', { status: body.title + ': ' + body.link },  function(error, tweet, response) {
-    if(error) {
-      mailError('Error posting to Twitter', error);
-      return;
-    }
-    
+  client.post('statuses/update', { status: `${ body.title }: ${ body.link }` },  (error, tweet, response) => {
+    if(error) return mailError('Error posting to Twitter', error);
     console.log('Twitter: ', response.statusMessage);
   });
 };
