@@ -1,6 +1,7 @@
 // Import Angular modules
 import { Component } from '@angular/core';
 import { Router } from '@angular/router';
+import { Store } from '@ngrx/store';
 
 // Import services
 import { HttpService } from '../services/http.service';
@@ -20,12 +21,18 @@ import { HttpService } from '../services/http.service';
 
 export class LoginComponent {
   public notification: Object;
+  public session: Object;
   public user = {
     email: null,
     password: null
   };
 
-  constructor(private http: HttpService, private router: Router) {}
+  constructor(private http: HttpService, private router: Router, private store: Store<any>) {
+    store.select('session').subscribe(session => this.session = session);
+    
+    if (this.session)
+      this.router.navigate(['/config']);
+  }
   
   public loginUser() {
     if (!this.user.email)
@@ -35,8 +42,11 @@ export class LoginComponent {
       return this.notification = { type: 'error', payload: 'Please enter a valid password' };
     
     this.http.loginUser(this.user).subscribe(
-      res => this.router.navigate(['/config']),
-      err => this.notification = { type: 'error', payload: err }
+      session => {
+        this.store.dispatch({ type: 'SET_SESSION', payload: JSON.parse(session['_body']) });
+        this.router.navigate(['/config']);
+      },
+      err => this.notification = { type: 'error', payload: JSON.parse(err._body).message }
     );
   }
 }
