@@ -1,6 +1,7 @@
 // Import Angular modules
 import { Component } from '@angular/core';
 import { Router } from '@angular/router';
+import { Store } from '@ngrx/store';
 
 // Import services
 import { HttpService } from '../services/http.service';
@@ -20,6 +21,7 @@ import { HttpService } from '../services/http.service';
 
 export class RegisterComponent {
   public notification: Object;
+  public session: Object;
   public user = {
     first_name: null,
     last_name: null,
@@ -27,24 +29,32 @@ export class RegisterComponent {
     password: null
   };
 
-  constructor(private http: HttpService, private router: Router) {}
-  
+  constructor(private http: HttpService, private router: Router, private store: Store<any>) {
+    store.select('session').subscribe(session => this.session = session);
+
+    if (this.session)
+      this.router.navigate(['/config']);
+  }
+
   public registerUser() {
     if (!this.user.first_name)
       return this.notification = { type: 'error', payload: 'Please enter your first name' };
-    
+
     if (!this.user.last_name)
       return this.notification = { type: 'error', payload: 'Please enter your last name' };
-    
+
     if (!this.user.email)
       return this.notification = { type: 'error', payload: 'Please enter your email' };
-    
+
     if (!this.user.password)
       return this.notification = { type: 'error', payload: 'Please enter a valid password' };
-    
+
     this.http.registerUser(this.user).subscribe(
-      res => this.router.navigate(['/config']),
-      err => this.notification = { type: 'error', payload: err }
+      session => {
+        this.store.dispatch({ type: 'SET_SESSION', payload: JSON.parse(session['_body']) });
+        this.router.navigate(['/config']);
+      },
+      err => this.notification = { type: 'error', payload: JSON.parse(err._body).message }
     );
   }
 }
